Fix upvote/downvote functionality
The merge of PR #187 broke the up- and downvote functionality because
now the simple-name of the "related item" class is passed to the server.
This extends ModelUrlMapping to not only handle the translated names but
also the simple class names. The upvote/downvote URLs are only called
from JavaScript (thus they are not seen by any user, except for those
who are looking at the requests in their browser's development tools).