Clear the "login required" message when the user is already logged in
In case of LDAP SSO the user is trying to access Mamute (while still not
authenticated via LDAP SSO). The user is then redirected to the login
where LDAP SSO is enabled. When the request reaches AuthController the
user is already logged in and redirected to the page where he/she came
from. Unfortunately the "you must be logged in" message was still shown
in this case. Fix this by clearing the message in such cases.